applied HOC to members + chapters + admin pages
